ig_scraper/requirements.txt
playwright
python-dotenv
streamlit
pandas

httpx
